+++ src/Commands/NewPackage.php
@@ -2,10 +2,13 @@
 
 namespace JeroenG\Packager\Commands;
 
-use JeroenG\Packager\Conveyor;
-use JeroenG\Packager\Wrapping;
 use Illuminate\Console\Command;
+use Illuminate\Contracts\Validation\Validator as ValidatorInterface;
+use Illuminate\Support\Facades\Validator;
+use JeroenG\Packager\Conveyor;
 use JeroenG\Packager\ProgressBar;
+use JeroenG\Packager\ValidationRules\ValidClassName;
+use JeroenG\Packager\Wrapping;
 
 /**
  * Create a brand new package.
@@ -20,7 +23,7 @@ class NewPackage extends Command
      * The name and signature of the console command.
      * @var string
      */
-    protected $signature = 'packager:new {vendor} {name} {--i}';
+    protected $signature = 'packager:new {vendor?} {name?} {--i} {--skeleton=}';
 
     /**
      * The console command description.
@@ -62,13 +65,31 @@ class NewPackage extends Command
         // Start the progress bar
         $this->startProgressBar(6);
 
+        $vendor = $this->argument('vendor');
+        $name = $this->argument('name');
+
+        if (stripos($vendor, '/') > 0) {
+            $part = explode('/', $vendor);
+            $vendor = $part[0];
+            $name = $part[1];
+        }
+
         // Defining vendor/package, optionally defined interactively
         if ($this->option('i')) {
-            $this->conveyor->vendor($this->ask('What will be the vendor name?', $this->argument('vendor')));
-            $this->conveyor->package($this->ask('What will be the package name?', $this->argument('name')));
+            $this->conveyor->vendor($this->ask('What will be the vendor name?', $vendor));
+            $this->conveyor->package($this->ask('What will be the package name?', $name));
         } else {
-            $this->conveyor->vendor($this->argument('vendor'));
-            $this->conveyor->package($this->argument('name'));
+            $this->conveyor->vendor($vendor);
+            $this->conveyor->package($name);
+        }
+
+        // Validate the vendor and package names
+        $validator = $this->validateInput($this->conveyor->vendor(), $this->conveyor->package());
+
+        if ($validator->fails()) {
+            $this->showErrors($validator);
+
+            return 1;
         }
 
         // Start creating the package
@@ -88,7 +109,11 @@ class NewPackage extends Command
 
         // Get the packager package skeleton
         $this->info('Downloading skeleton...');
-        $this->conveyor->downloadSkeleton();
+        if ($this->option('i')) {
+            $this->conveyor->downloadSkeleton($this->ask('What package skeleton would you like to use?', $this->option('skeleton') ?? config('packager.skeleton')));
+        } else {
+            $this->conveyor->downloadSkeleton($this->option('skeleton') ?? null);
+        }
         $manifest = (file_exists($this->conveyor->packagePath().'/rewriteRules.php')) ? $this->conveyor->packagePath().'/rewriteRules.php' : null;
         $this->conveyor->renameFiles($manifest);
         $this->makeProgress();
@@ -101,8 +126,8 @@ class NewPackage extends Command
             ':lc:vendor',
             ':lc:package',
         ], [
-            $this->conveyor->vendor(),
-            $this->conveyor->package(),
+            $this->conveyor->vendorStudly(),
+            $this->conveyor->packageStudly(),
             strtolower($this->conveyor->vendor()),
             strtolower($this->conveyor->package()),
         ]);
@@ -169,5 +194,22 @@ class NewPackage extends Command
             $description,
             $license,
         ]);
+    }
+
+    private function validateInput(string $vendor, string $name)
+    {
+        return Validator::make(compact('vendor', 'name'), [
+            'vendor' => new ValidClassName,
+            'name' => new ValidClassName,
+        ]);
+    }
+
+    private function showErrors(ValidatorInterface $validator)
+    {
+        $this->info('Package was not created. Please choose a valid name.');
+
+        foreach ($validator->errors()->all() as $error) {
+            $this->error($error);
+        }
     }
 }